Update dependencies.
It looks like it's really hard to make a stream that messes with PDFBox
now. I had to disable some tests, because I couldn't cause PDFBox to
throw an exception when reading a single page. (It can throw an
exception when loading an entire document, but I can't make it throw an
exception just on the second page.)